requirements.txt
asgiref==3.7.2
Django==4.2.3
djangorestframework==3.14.0
lorem==0.1.1
pytz==2023.3
sqlparse==0.4.4
tqdm==4.65.0
